updating relative position using translation and rotation classes
